fix connector metadata for upload collection
